+++ src/BlackJack/Form1.cs
@@ -10,10 +10,10 @@ using System.Windows.Forms;
 
 namespace BlackJack
 {
-    public partial class BlackJack : Form 
-    {  
-         
-         
+    public partial class BlackJack : Form
+    {
+
+
         private static int startXPos = 259; //Стартовая позиция карт игрока (можно(нужно) менять)
         private static int playerCardYPos = 200; // y-координата карт игрока 
         private int playerCardXPos = startXPos;// Позиция карт игрока, в который будт отрисоваться каты (Егор не трогай!)
@@ -22,17 +22,21 @@ namespace BlackJack
         private static int dealerCardYPos = 18; // y-координата карт дилера 
         private int dealerCardXPos = dealerStartXPos;// Позиция карт дилера, в который будт отрисоваться каты (Егор не трогай!)
 
-        private string[] PlayerCardUpdate { get { string[] param = textboxPlayerCards.Text.Split(new char[] {' '}, StringSplitOptions.RemoveEmptyEntries);return param; } }//Заносим в массив все карты игрока
+        private string[] PlayerCardUpdate { get { string[] param = textboxPlayerCards.Text.Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries); return param; } }//Заносим в массив все карты игрока
         private string[] DealerCardUpdate { get { string[] param = D1.Text.Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries); return param; } }//Заносим в массив все карты игрока
 
         public List<PictureBox> playerCardsToDisplay; // Объявляем лист  карт для игрока
         public List<PictureBox> dealerCardsToDisplay; // Объявляем лист  карт для дилера
+        
+
+        public int totaldealerwithbolckcard = 0;//очки дилера с блокированной картой
 
         public BlackJack()
-        { 
+        {
             InitializeComponent();
             playerCardsToDisplay = new List<PictureBox>(); // создаем Лист , чтобы заносить рисунки карт игрока
             dealerCardsToDisplay = new List<PictureBox>(); // Создаем Лист , чтобы заносить рисунки карт дилера
+            
         }
 
         Game blackjack = new Game();
@@ -40,39 +44,66 @@ namespace BlackJack
         {
 
         }
-         
+
+        /// <summary>
+        /// Костыль что бы получить очки только первой карты дилера 
+        /// </summary>
+        /// <returns></returns>
+        public int gettotaldealerwithbolckcard()
+        {
+            return blackjack.GetTotalDealer() - Game.DealerSecondCard.GetValue(); ;
+        }
+
         private void newGame_Click(object sender, EventArgs e)
         {
-            RemoveCards(playerCardsToDisplay); RemoveCards(dealerCardsToDisplay); //Очишаем лист с картами игрка  и дилера
-            playerCardXPos = startXPos; dealerCardXPos = dealerStartXPos;//!!Востанавливаем позицию появления карт игрока и дилера
+            PlayerCardsRemoveAndBackToStartPosition(); DealerCardsRemoveAndBackToStartPosition(); //Очишаем лист с картами игрка  и дилера !!Востанавливаем позицию появления карт игрока и дилера
+   
+            // Показать все элементы после старта игры
+            buttonStatistics.Visible = true;
+            label1.Visible = true;
+            buttonStatistics.Visible = true;
+            label2.Visible = true;
+            dealerScore.Visible = true;
+            label4.Visible = true;
+            dealerScore.Visible = true;
+            label6.Visible = true;
+            info.Visible = true;
+            playerScore.Visible = true;
+            labelTitle.Visible = false;
+            labelSubtitle.Visible = false;
+
 
             // очищаем label с тузом  
-            labelHasA.Text = ""; 
-            blackjack.EndGame();   
-            blackjack.GameStart(); 
+            labelHasA.Text = "";
+            blackjack.EndGame();
+            blackjack.GameStart();
+
+
 
             buttonSplit.Visible = false;
+
             if (blackjack.SplitCheck())
             {
                 buttonSplit.Visible = true;
             }
-                
+           
             //выводим на форму первые 2 карты дилера и игрока 
-            D1.Text = blackjack.DealerCards;                    
+            D1.Text = blackjack.DealerCards;
             textboxPlayerCards.Text = blackjack.PlayerCards;
             // их очки
             playerScore.Text = blackjack.GetTotalPlayer().ToString();
-            dealerScore.Text = blackjack.GetTotalDealer().ToString();
-           
+            //Теперь очки только перывой карты дилера 
+            dealerScore.Text = gettotaldealerwithbolckcard().ToString();
 
-            HitButton.Visible = true;  
+
+            HitButton.Visible = true;
             StandButton.Visible = true;
 
             if (blackjack.PlayerHasA()) //если в картах игрока имеется туз
             {
                 labelHasA.Text = "A(1 или 11)"; //выводим соотв текст 
             }
-             
+
             info.Text = "Берите карты или остановитесь"; //проверка на 21 после начала игры 
             if (blackjack.GetTotalPlayer() == 21) HitButton.Visible = false;
 
@@ -80,70 +111,77 @@ namespace BlackJack
             PlayerCardDraw();
             DealerCardDraw();
 
+            DrawBlockedDealer();
 
-        } 
+        }
 
         private void HitButton_Click(object sender, EventArgs e)
         {
-            // очищаем label с тузом  
+            // очищаем label с тузом   
             labelHasA.Text = "";
 
-             blackjack.Hit(Game.PlayerCardsList); 
-             
+            blackjack.Hit(Game.PlayerCardsList);
+
             if (blackjack.PlayerHasA()) //если в картах игрока имеется туз
             {
                 labelHasA.Text = "A(1 или 11)"; //выводим соотв текст 
             }
-             
-            if (buttonSplit.Visible == true) 
+
+            if (buttonSplit.Visible == true)
                 buttonSplit.Visible = false;
 
             if (blackjack.IsOverflow()) //если перебор у игрока 
-            { 
+            {
                 // присуждаем победу дилеру, заканчиваем игру 
                 //dealersWinsText.Text = Statistic.GetDealerWins().ToString(); 
-                info.Text = "Победа дилера! Перебор у игрока"; 
+                info.Text = "Победа дилера! Перебор у игрока";
                 HitButton.Visible = false;
-                StandButton.Visible = false;  
-            }  
-             // записываем новую карту 
+                StandButton.Visible = false;
+                //Проводим операции над картами дилера 
+                DealerCardsRemoveAndBackToStartPosition();
+                DealerCardDraw();
+                dealerScore.Text = blackjack.GetTotalDealer().ToString();
+            }
+            // записываем новую карту 
             playerScore.Text = blackjack.GetTotalPlayer().ToString();
-            
+
             textboxPlayerCards.Text = blackjack.PlayerCards;
             if (blackjack.GetTotalPlayer() == 21) HitButton.Visible = false;
 
             // При хите отрисововаем карты
-            RemoveCards(playerCardsToDisplay);//Очишаем лист с картами игрка 
-            playerCardXPos = startXPos;//!!Востанавливаем позицию появления карт игрока
+            PlayerCardsRemoveAndBackToStartPosition();//Очишаем лист с картами игрка !!Востанавливаем позицию появления карт игрока
             PlayerCardDraw();// Рисуем
         }
-         
+
         private void StandButton_Click(object sender, EventArgs e)
-        { 
-            blackjack.DealerMove();  
+        {
+            blackjack.DealerMove();
             //выводим очки и карты дилера
             dealerScore.Text = blackjack.GetTotalDealer().ToString();
             D1.Text = blackjack.DealerCards;
             //сообщение о конце игры
-            info.Text = blackjack.EndMessage();   
+            info.Text = blackjack.EndMessage();
             // очищаем label с тузом  
-            labelHasA.Text = "";   
+            labelHasA.Text = "";
             // скрываем кнопки  
-            HitButton.Visible = false;   
+            HitButton.Visible = false;
             StandButton.Visible = false;
 
             // При стенде отрисововаем карты
-            RemoveCards(dealerCardsToDisplay);//Очишаем лист с картами дилера
-            dealerCardXPos = dealerStartXPos;//!!Востанавливаем позицию появления карт дилера
+            DealerCardsRemoveAndBackToStartPosition();//Очишаем лист с картами дилера//!!Востанавливаем позицию появления карт дилера
             DealerCardDraw(); // Рисуем
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
+
             Split f = new Split();
-            f.Owner = this; 
+            f.Owner = this;
             f.ShowDialog();
             buttonSplit.Visible = false;
+            //DealerCardsRemoveAndBackToStartPosition(); // Егор если сделашь, чтоб при сплите закрывалась основная форма , то убери коменты   ♥#VS Chat 20!8♥
+            //DealerCardDraw(); // 
+
         }
 
         private void buttonStatistics_Click(object sender, EventArgs e)
@@ -153,13 +191,13 @@ namespace BlackJack
 
         }
 
-       
+
         /// <summary>
         /// Функция для отрисовки карт
         /// </summary>
         /// <param name="card">передаем карту, которую хотим отрисовать</param>
         private void DrawPlayer(String card)// 
-        { 
+        {
             playerCardXPos += 35;//Расстояние между картами 
             PictureBox newCard = new PictureBox();//Создаем объект класса PictureBox
             Image img = Image.FromFile("../../CardsSprites/" + card + ".png"); //берем саму картинку из папки проекта , пердворительно узнав ее названия
@@ -170,8 +208,8 @@ namespace BlackJack
             this.Controls.Add(newCard);//хз
             newCard.BringToFront();// на верх всего
             playerCardsToDisplay.Add(newCard);//добавляем в лист
-        } 
-      
+        }
+
 
         /// <summary>
         /// Функция для очистки карт с формы
@@ -192,7 +230,7 @@ namespace BlackJack
         {
             string[] cardUpdate = PlayerCardUpdate;// заносим в данные в массив
             for (int i = 0; i < cardUpdate.Length; i = i + 2)// циклом берм из массива имена нужных карт и рисуем их
-            { 
+            {
                 string card;
                 card = cardUpdate[i] + cardUpdate[i + 1];
                 DrawPlayer(card);
@@ -230,6 +268,42 @@ namespace BlackJack
                 DrawDealer(card);
             }
         }
-    }       
-}  
-   
+
+      
+        /// <summary>
+        /// Ну ты понял ;-) чтою вместо двух строк кода, одну писать оптимизация
+        /// </summary>
+        private void PlayerCardsRemoveAndBackToStartPosition()
+        {
+            RemoveCards(playerCardsToDisplay);//Очишаем лист с картами игрка 
+            playerCardXPos = startXPos;//!!Востанавливаем позицию появления карт игрока
+        }
+       
+        /// <summary>
+        /// Ну ты понял ;-) чтою вместо двух строк кода, одну писать оптимизация
+        /// </summary>
+        private void DealerCardsRemoveAndBackToStartPosition()
+        {
+            RemoveCards(dealerCardsToDisplay);
+            dealerCardXPos = dealerStartXPos;
+        }
+
+        /// <summary>
+        /// Рисуем Закрытую карту 
+        /// </summary>
+        private void DrawBlockedDealer()
+        {
+                dealerCardXPos -= 35;//Расстояние между картами 
+                PictureBox newCard = new PictureBox();//Создаем объект класса PictureBox
+                Image img = Image.FromFile("../../CardsSprites/b2fv.png");//берем саму картинку из папки проекта , пердворительно узнав ее названия
+                newCard.Image = img;//присваиваем картинку
+                newCard.Location = new System.Drawing.Point(329, 18);
+                newCard.Name = "newCard";
+                newCard.Size = new System.Drawing.Size(72, 99);// Размер картинки
+                this.Controls.Add(newCard);
+                newCard.BringToFront();// на верх всего
+                
+        }
+    }
+}
+   //                                               ♥ We Love Anime♥
